scraper_service/pyproject.toml
[tool.poetry]
name = "rent-service-scraper"
version = "0.1.0"
description = "Scraper service for rent monitoring"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
schedule = "^1.2.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
alembic = "^1.13.1"
redis = "^5.0.1"
free-proxy = "^1.1.3"
faststream = {extras = ["rabbit"], version = "^0.5.34"}
pydantic-settings = "^2.2.1"

httpx = "^0.28.1"
telethon = "^1.28.5"
together = "^0.1.7"
aiohttp = "^3.8.5"
aiofiles = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.9.1"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"

[tool.poe.tasks.format]
shell = "black src && isort src && flake8 src --format='scraper_service/%(path)s:%(row)d:%(col)d: %(code)s %(text)s'"

[tool.flake8]
max-line-length = 120
max-complexity = 15
ignore = "E501, W503, E722"
per-file-ignores = "__init__.py:F401"

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2

[tool.pylint.format]
max-line-length = "120"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 